refactor(registry): split rekor verifier helpers
